% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_comp_mse.R
\name{comp_mse}
\alias{comp_mse}
\title{Compute MSE of Parameters}
\usage{
comp_mse(
  N,
  iter,
  I,
  mu.person = c(0, 0),
  mu.item = c(1, 0, 1, 1),
  meanlog.sigma2 = log(0.2),
  cov.m.person = matrix(c(1, 0.5, 0.5, 1), ncol = 2, byrow = TRUE),
  cov.m.item = matrix(c(0.2, 0, 0, 0, 0, 0.5, -0.35, -0.15, 0, -0.35, 0.5, 0.15, 0,
    -0.15, 0.15, 0.2), ncol = 4, byrow = TRUE),
  sdlog.sigma2 = 0.2,
  item.pars.m = NULL,
  cor2cov.item = FALSE,
  sd.item = NULL,
  XG = 3000,
  burnin = 20,
  mse.seed = NULL,
  keep.err.dat = TRUE
)
}
\arguments{
\item{N}{Integer. The sample size.}

\item{iter}{Integer. The number of iteration or the number of data sets.}

\item{I}{Integer. The test length.}

\item{mu.person}{Numeric vector. Means of theta and zeta}

\item{mu.item}{Numeric vector. Means of alpha, beta, phi, and lambda}

\item{meanlog.sigma2}{Numeric. The meanlog of sigma2.}

\item{cov.m.person}{Matrix. The covariance matrix of theat and zeta}

\item{cov.m.item}{Matrix. The covariance matrix of of alpha, beta, phi, and lambda}

\item{sdlog.sigma2}{Numeric. The sdlog of sigma2}

\item{item.pars.m}{Matrix. (optional) A Matrix containing item parameters.}

\item{cor2cov.item}{Logical. Whether a correlation matrix instead of covariance matrix is supplied}

\item{sd.item}{Numeric vector. (optional) The standard deviations of alpha, beta, phi, and lambda}
}
\value{
A list of containing:
\describe{
  \item{mse.theta}{Numeric. The pooled MSE of the theta parameters.}
  \item{mse.alpha}{Numeric. The pooled MSE of the alpha paramters.}
  \item{mse.beta}{Numeric. The pooled MSE of the beta paramters.}
  \item{mse.phi}{Numeric. The pooled MSE of the phi paramters.}
  \item{mse.lambda}{Numeric. The pooled MSE of the lambda paramters.}
  \item{mse.sigma2}{Numeric. The pooled MSE of the sigma2 paramters.}
  \item{conv.rate}{Data Frame. The Rhat convergence rate per iteration (rows)
  by parameter blocks (columns).}
}
}
\description{
The function computes the mean squared errors of estimated parameters based on
simulated data under the the Joint Hierarchical Model using a 2-pl normal
ogive model for response accuracy and a 3-pl log-normal model for response
time.
}
\examples{
 \dontrun{
test = comp_mse(
          iter = 5,
          N = 100,
          I = 10,
          mu.person = c(0,0),
          mu.item = c(1,0,1,1),
          meanlog.sigma2 = log(1),
          cov.m.person = matrix(c(1,0,
                                  0,1), ncol = 2, byrow = TRUE),
          cov.m.item = matrix(c(1, 0, 0, 0,
                                0, 1, 0, 0,
                                0, 0, 1, 0,
                                0, 0, 0, 1), ncol =  4, byrow = TRUE),
          sd.item         = c(.2, 1, .2, .5),
          cor2cov.item    = TRUE,
          sdlog.sigma2 = 0.2,
          XG = 2000,
          keep.err.dat = TRUE)
}
}
