% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_design_conditions.R
\name{design_conditions}
\alias{design_conditions}
\title{Generate the Design Conditions}
\usage{
design_conditions()
}
\value{
A list containing  parameters
\describe{
  \item{out.par}{Character. The name of the target parameter for the threshold.}
  \item{thresh}{Numeric. The desired threshold of the target parameter to be achieved.}
  \item{I}{Integer. The test length.}
  \item{cov.m.person}{Matrix. The covariance matrix of theat and zeta}
  \item{mu.item}{Numeric vector. Means of alpha, beta, phi, and lambda.}
  \item{cov.m.item}{Matrix. The covariance matrix of of alpha, beta, phi, and lambda.}
}
}
\description{
The function generates a list with specified design conditions for all
manipulated parameters of the sample size estimation function.
}
\examples{
 \dontrun{
design_conditions()
}

}
