% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_geweke_LNIRT.R
\name{geweke.LNIRT}
\alias{geweke.LNIRT}
\title{Geweke Convergence for LNIRT objects}
\usage{
geweke.LNIRT(object, z.cut = 1.96)
}
\arguments{
\item{object}{List. A fitted LNIRT object.}

\item{z.cut}{Numeric. The cutoff z-value (two-tailed) for rejecting convergence.}
}
\value{
A list containing:
\describe{
  \item{D1}{Numeric. Proportion with |z| ≤ z.cut for hyperparameters.}
  \item{D2.item}{Numeric. Proportion with |z| ≤ z.cut for item-level parameters.}
  \item{D2.person}{Numeric. Proportion with |z| ≤ z.cut for person-level parameters.}
  \item{D3}{Numeric. Proportion with |z| ≤ z.cut for item covariance elements.}
}
}
\description{
The function calculates Geweke's convergence diagnostics (Geweke, 1992) for
all parameters from a LNIRT objects.
}
