% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_item_par.R
\name{item.par}
\alias{item.par}
\title{Generate Item parameters}
\usage{
item.par(
  I,
  mu.item = c(1, 0, 4, 0),
  cov.m.item = matrix(c(0.2, 0, 0, 0, 0, 0.5, -0.35, -0.15, 0, -0.35, 0.5, 0.15, 0,
    -0.15, 0.15, 0.2), ncol = 4, byrow = TRUE),
  meanlog.sigma2 = log(0.3),
  sdlog.sigma2 = 0.2,
  cor2cov.item = FALSE,
  sd.item = NULL,
  item.seed = NULL
)
}
\arguments{
\item{I}{Integer. The test length or number of items.}

\item{mu.item}{Numeric vector. Means of alpha, beta, phi, and lambda}

\item{cov.m.item}{Matrix. The covariance matrix of of alpha, beta, phi, and lambda}

\item{meanlog.sigma2}{Numeric. The meanlog of sigma2}

\item{sdlog.sigma2}{Numeric. The sdlog of sigma2}

\item{cor2cov.item}{Logical. Whether a correlation matrix instead of covariance matrix is supplied}

\item{sd.item}{Numeric vector. (optional) The standard deviations of alpha, beta, phi, and lambda}

\item{item.seed}{Integer. (optional) Seed for drawing samples from distributions.}
}
\value{
A data frame containing item parameters
\describe{
  \column{alpha}{Numeric vector. The values of the item discrimination parameter.}
  \column{beta}{Numeric vector. The values of the item difficulty parameter.}
  \column{phi}{Numeric vector. The values of the time discrimination parameter.}
  \column{lambda}{Numeric vector. The values of the time intensity parameter.}
  \column{sigma2}{Numeric vector. The values of the error variance parameter.}
}
}
\description{
The function generates a data frame with item parameters from a truncated
multivariate normal distribution for alpha, beta, phi, lambda, and from a
lognormal distribution for sigma2.
}
\examples{
 \dontrun{

}
}
