% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_optim_sample.R
\name{optim.sample}
\alias{optim.sample}
\title{Optimize for Sample Size}
\usage{
optim.sample(
  FUN = comp_mse,
  thresh,
  lb,
  ub,
  tol,
  out.par = "mse.alpha",
  N = NULL,
  iter = 1,
  I = 20,
  mu.person = c(0, 0),
  mu.item = c(1, 0, 4, 0),
  meanlog.sigma2 = log(0.3),
  cov.m.person = matrix(c(1, 0.5, 0.5, 1), ncol = 2, byrow = TRUE),
  cov.m.item = matrix(c(0.2, 0, 0, 0, 0, 0.5, -0.35, -0.15, 0, -0.35, 0.5, 0.15, 0,
    -0.15, 0.15, 0.2), ncol = 4, byrow = TRUE),
  sdlog.sigma2 = 0.2,
  scale = TRUE,
  random.item = TRUE,
  item.pars.m = NULL,
  item.seed = 12345,
  person.seed = NULL,
  cor2cov.item = FALSE,
  sd.item = NULL,
  n.cores = NULL
)
}
\arguments{
\item{FUN}{Function. The function to calcuate the target parameter.}

\item{thresh}{Numeric. The desired threshold of the target parameter to be achieved.}

\item{lb}{Integer. The lower bound of the sample size.}

\item{ub}{Integer. The upper bound of the sample size.}

\item{tol}{Numeric. The tolerance in objective allowed for stopping the optimizer.}

\item{out.par}{Character. The name of the target parameter for the threshold.}

\item{N}{NULL. Place holder for the sample size in the optimizer.}

\item{iter}{Integer. The number of iteration or the number of data sets.}

\item{I}{Integer. The test length.}

\item{mu.person}{Numeric vector. Means of theta and zeta}

\item{mu.item}{Numeric vector. Means of alpha, beta, phi, and lambda}

\item{meanlog.sigma2}{Numeric. The meanlog of sigma2.}

\item{cov.m.person}{Matrix. The covariance matrix of theat and zeta}

\item{cov.m.item}{Matrix. The covariance matrix of of alpha, beta, phi, and lambda}

\item{sdlog.sigma2}{Numeric. The sdlog of sigma2}

\item{scale}{Logical. Weather the item and person parameters are scaled.}

\item{random.item}{Logical. Weather the item parameters are sampled.}

\item{item.pars.m}{Matrix. (optional) A Matrix containing item parameters.}

\item{item.seed}{Integer. (optional) Seed for drawing samples from item parameter distributions.}

\item{person.seed}{Integer. (optional) Seed for drawing samples from person parameter distributions.}

\item{cor2cov.item}{Logical. Whether a correlation matrix instead of covariance matrix is supplied}

\item{sd.item}{Numeric vector. (optional) The standard deviations of alpha, beta, phi, and lambda}

\item{n.cores}{Integer. (optional) The number of cores for parallel computation.}
}
\value{
A list of containing:
\describe{
  \item{N.best}{Integer. The minimum sample size for the closest result to the threshold.}
  \item{res.best}{Numeric. The result of the target parameter closest to the threshold.}
  \item{reps}{Integer. Number of optimization repetitions.}
  \item{track.res}{Data Frame. The results around the threshold per repetition.}
  \item{track.N}{Data Frame. The sample size in each repetition.}
  \item{mse.sigma2}{Numeric. The pooled MSE of the sigma2 paramters.}
  \item{track.conv}{List. The Rhat convergence rate per repetition.}
}
}
\description{
The function optimizes for the minimum sample size to reach the threshold of mean
squared errors (MSE) of estimated parameters. MSE are based on simulated data
under the the Joint Hierarchical Model using a 2-pl normal ogive model for
response accuracy and a 3-pl log-normal model for response time.
}
