% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_person_par.R
\name{person.par}
\alias{person.par}
\title{Generate Person Parameters}
\usage{
person.par(
  N,
  mu.person = c(0, 0),
  cov.m.person = matrix(c(1, 0.5, 0.5, 1), ncol = 2, byrow = TRUE),
  person.seed = NULL
)
}
\arguments{
\item{N}{Integer. The sample size or number of rows.}

\item{mu.person}{Numeric vector. Means of theta and zeta}

\item{cov.m.person}{Matrix. The covariance matrix of theat and zeta}
}
\value{
A data frame containing person parameters
\describe{
  \column{theta}{Numeric vector. The values of theta.}
  \column{zeta}{Numeric vector. The values of zeta,}
}
}
\description{
The function generates a data frame with person parameters from a
multivariate normal distribution for theta and zeta.
}
\examples{
 \dontrun{
test.data <- person.par(N = 50,
                          mu.person = c(0,0),
                         cov.m.person = matrix(c(1,.5,
                                                 .5,1), ncol = 2, byrow = TRUE),
                         person.seed = 123)
}

}
