% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_rhat_LNIRT.R
\name{rhat_LNIRT}
\alias{rhat_LNIRT}
\title{Rhat for LNIRT objects}
\usage{
rhat_LNIRT(object.list, chains = 4, cutoff = 1.01)
}
\arguments{
\item{object.list}{List. A list containing fitted LNIRT objects.}

\item{chains}{Integer. The number of chains to consider.}

\item{cutoff}{Numeric. The Rhat cutoff value for convergence.}
}
\value{
A list containing:
\describe{
  \item{value}{List. R-hat values by block}
  \item{convergence}{List. Binary indicators (1 if R-hat < \code{cutoff}, else 0) by block}
  \item{rate}{List. Mean convergence rate (proportion below \code{cutoff}) by block}
}
}
\description{
The function calculates Rhat convergence diagnostics (Vehtari et al., 2021) for
all parameters from a list of LNIRT objects.
}
