% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_scale_M.R
\name{scale_M}
\alias{scale_M}
\title{Scale Item and Person Matrices}
\usage{
scale_M(item.pars, person.pars, re.scale = FALSE, c.alpha = NULL, c.phi = NULL)
}
\arguments{
\item{item.pars}{Data frame or Matrix. The item parameters.}

\item{person.pars}{Data frame or Matrix. The person parameters.}
}
\value{
A list containing:
\describe{
  \item{items.pars.scaled}{Matrix. The scaled item parameters.}
  \item{person.pars.scaled}{Matrix. The scaled person parameters.}
  \item{c.alpha}{Numeric. The scaling constant of alpha.}
  \item{c.phi}{Numeric. The scaling constant of phi}
}
}
\description{
The function scales the matrices containing item and person parameters by setting the
geometric means of alpha and phi being equal to 1, to align with the LNIRT package.
}
