% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sim_jhm_data.R
\name{sim.jhm.data}
\alias{sim.jhm.data}
\title{Simulate Data under Joint Hierarchical Model}
\usage{
sim.jhm.data(
  iter,
  N,
  K,
  mu.person = c(0, 0),
  mu.item = c(1, 0, 4, 0),
  meanlog.sigma2 = log(0.3),
  cov.m.person = matrix(c(1, 0.5, 0.5, 1), ncol = 2, byrow = TRUE),
  cov.m.item = matrix(c(0.2, 0, 0, 0, 0, 0.5, -0.35, -0.15, 0, -0.35, 0.5, 0.15, 0,
    -0.15, 0.15, 0.2), ncol = 4, byrow = TRUE),
  sdlog.sigma2 = 0.2,
  item.pars.m = NULL,
  cor2cov.item = FALSE,
  sd.item = NULL,
  scale = TRUE
)
}
\arguments{
\item{iter}{Integer. The number of iteration or the number of data sets.}

\item{N}{Integer. The sample size.}

\item{K}{Integer. The test length.}

\item{mu.person}{Numeric vector. Means of theta and zeta}

\item{mu.item}{Numeric vector. Means of alpha, beta, phi, and lambda}

\item{meanlog.sigma2}{Numeric. The meanlog of sigma2.}

\item{cov.m.person}{Matrix. The covariance matrix of theta and zeta}

\item{cov.m.item}{Matrix. The covariance matrix of of alpha, beta, phi, and lambda}

\item{sdlog.sigma2}{Numeric. The sdlog of sigma2}

\item{item.pars.m}{Matrix. (optional) A Matrix containing item parameters.}

\item{cor2cov.item}{Logical. Whether a correlation matrix instead of covariance matrix is supplied}

\item{sd.item}{Numeric vector. (optional) The standard deviations of alpha, beta, phi, and lambda}

\item{scale}{Logical. Whether to rescale item parameters.}
}
\value{
A list of length `iter` containing:
\describe{
  \item{long.data}{Data Frame. The simulated data in long format.}
  \item{time.data}{Data Frame. The simulated response time data.}
  \item{response.data}{Data Frame. The simulated response accuracy data.}
  \item{person.par}{Data Frame. The person parameters.}
  \item{item.par}{Data Frame. The item parameters.}
}
}
\description{
The function simulates data under the the Joint Hierarchical Model using a
2-pl normal ogive model for response accuracy and a 3-pl log-normal model for response
time.
}
\examples{
 \dontrun{
test.data <- sim.jhm.data(iter = 2,
                         N = 100,
                         K = 10,
                         mu.person = c(0,0),
                         mu.item = c(1,0,1,0),
                         meanlog.sigma2 = log(.3),
                         cov.m.person = matrix(c(1,0.5,
                                                 0.5,1), ncol = 2, byrow = TRUE),
                         cov.m.item = matrix(c(.2, 0, 0, 0,
                                               0, .5, 0, 0,
                                               0, 0, .5, 0,
                                               0, 0, 0, .2), ncol =  4, byrow = TRUE),
                         sdlog.sigma2 = 0.2
                         )
}

}
