% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{sspLNIRT-package}
\alias{sspLNIRT}
\alias{sspLNIRT-package}
\title{sspLNIRT: SampleSize Planning for the Joint Hierarchical Model}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Sample size planning tools for the Joint Hierarchical Model using a 2-PL normal ogive model for response accuracy and a 3-PL log-normal model for response time. Precomputed results can be accessed instantly, custom designs can be computed, and diagnostics can be used to interpret results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/anonymous-peer-2026/sspLNIRT}
  \item \url{https://anonymous-peer-2026.github.io/sspLNIRT/}
  \item Report bugs at \url{https://github.com/anonymous-peer-2026/sspLNIRT/issues}
}

}
\author{
\strong{Maintainer}: Anonymous Anonymous \email{anonymous@example.com}

}
\keyword{internal}
