% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sspLNIRT.data}
\alias{sspLNIRT.data}
\title{Precomputed sample size planning results}
\format{
A nested list where each element is an \code{sspLNIRT.object}.
}
\usage{
sspLNIRT.data
}
\description{
A nested list of precomputed \code{optim_sample()} results across a
factorial grid of design conditions.
}
\keyword{datasets}
