% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_sspLNIRT.R
\name{sspLNIRT_scales}
\alias{sspLNIRT_scales}
\alias{scale_colour_sspLNIRT}
\alias{scale_fill_sspLNIRT}
\title{Discrete colour / fill scales for sspLNIRT plots}
\usage{
scale_colour_sspLNIRT(...)

scale_fill_sspLNIRT(...)
}
\arguments{
\item{...}{Passed to [ggplot2::discrete_scale()].}
}
\value{
A ggplot2 discrete scale.
}
\description{
Discrete ggplot2 scales matching [theme_sspLNIRT()]. The palette is a
grey-to-slate-blue ramp.
}
\examples{
\dontrun{
library(ggplot2)
ggplot(iris, aes(Sepal.Length, Sepal.Width, colour = Species)) +
  geom_point() +
  scale_colour_sspLNIRT() +
  theme_sspLNIRT()
}

}
