% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_summary.R
\name{summary.sspLNIRT.object}
\alias{summary.sspLNIRT.object}
\title{Summarize a sspLNIRT.object}
\usage{
\method{summary}{sspLNIRT.object}(object, ...)
}
\arguments{
\item{object}{A `sspLNIRT.object` produced by `optim_sample` or `comp_mse`.}

\item{...}{Additional arguments (unused).}
}
\value{
An object of class `summary.sspLNIRT.object`.
}
\description{
Provides a comprehensive summary of a `sspLNIRT.object`.
}
