% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_available_configs.R
\name{available_configs}
\alias{available_configs}
\title{Show available precomputed configurations}
\usage{
available_configs()
}
\value{
A tibble summarizing available design conditions
}
\description{
Show available precomputed configurations
}
