% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_precision.R
\name{plot_precision}
\alias{plot_precision}
\title{Plot the precision of item parameters}
\usage{
plot_precision(object, y.val = "mse", n.bins = NULL)
}
\arguments{
\item{object}{Object. Output from optim_sample() or comp_mse(). Must contain $err.dat.}

\item{y.val}{String. Either "mse" or "bias" for the y-axis.}

\item{n.bins}{Integer (optional). Number of bins used to aggregate by sim.val within par.}
}
\value{
A ggplot object
}
\description{
The function plots the mean squared error or bias across values of the true parameter.
}
\examples{
 \dontrun{
   plot_precision(object = test.optim.sample)
}

}
