% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{sspLNIRT-package}
\alias{sspLNIRT}
\alias{sspLNIRT-package}
\title{sspLNIRT: Sample Size Planning for Item Calibration using the Joint Hierarchical Model}
\description{
Sample size planning tools for the Joint Hierarchical Model using a 2-PL normal ogive model for response accuracy and a 3-PL log-normal model for response time, estimated via the LNIRT package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sebastian-lortz/sspLNIRT}
  \item \url{https://sebastian-lortz.github.io/sspLNIRT/}
  \item Report bugs at \url{https://github.com/sebastian-lortz/sspLNIRT/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Alexander Jens Lortz \email{s.a.j.lortz@gmail.com}

}
\keyword{internal}
