% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{sspLNIRT-package}
\alias{sspLNIRT}
\alias{sspLNIRT-package}
\title{sspLNIRT: Sample Size Planning for Joint Hierarchical Modelling using LNIRT}
\description{
Sample size planning tools for the Joint Hierarchical Model using a 2-PL normal ogive model for response accuracy and a 3-PL log-normal model for response time, estimated via the LNIRT package.
}
\author{
\strong{Maintainer}: Sebastian A. J. Lortz \email{s.a.j.lortz@gmail.com}

}
\keyword{internal}
